legacy/pyproject.toml
[project]
name = "fnirsi-ps-control"
dynamic = ["version"]
description = "PC control library and CLI for the FNIRSI DPS-150 regulated power supply"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.11"
keywords = ["fnirsi", "dps-150", "power-supply", "serial", "usb-cdc"]
classifiers = [
  "Development Status :: 2 - Pre-Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: System :: Hardware :: Hardware Drivers",
]

dependencies = [
  "pyserial>=3.5",
  "typer>=0.12",
  "rich>=13",
]

[project.optional-dependencies]
dev = [
  "pytest>=8",
  "pytest-cov>=5",
  "ruff>=0.4",
  "mypy>=1.10",
  "types-pyserial",
  "python-semantic-release>=9",
  "mkdocs>=1.6.1",
  "mkdocs-material>=9.7.6",
  "pymdown-extensions>=10.21.2",
]

[project.scripts]
fnirsi = "fnirsi_ps_control.cli:app"

[project.urls]
Homepage = "https://github.com/yourname/fnirsi_ps_control"
Repository = "https://github.com/yourname/fnirsi_ps_control"
"Issue Tracker" = "https://github.com/yourname/fnirsi_ps_control/issues"
Changelog = "https://github.com/yourname/fnirsi_ps_control/releases"

# ---------------------------------------------------------------------------
# Build
# ---------------------------------------------------------------------------
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "vcs"

[tool.hatch.version.raw-options]
# Follows PEP 440 / SemVer via git tags (e.g. v1.2.3).
# Local dev installs get a version like 0.1.0.dev3+gabcdef.
version_scheme = "python-simplified-semver"
local_scheme = "no-local-version"    # keeps PyPI-uploadable versions clean

[tool.hatch.build.hooks.vcs]
version-file = "src/fnirsi_ps_control/_version.py"

[tool.hatch.build.targets.wheel]
packages = ["src/fnirsi_ps_control"]

# ---------------------------------------------------------------------------
# Ruff – linting & formatting
# ---------------------------------------------------------------------------
[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "B", "C4", "ANN"]
ignore = ["ANN101", "ANN102"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["ANN"]

[tool.ruff.format]
quote-style = "double"

# ---------------------------------------------------------------------------
# Mypy
# ---------------------------------------------------------------------------
[tool.mypy]
python_version = "3.11"
strict = true
files = ["src"]

# ---------------------------------------------------------------------------
# Pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=fnirsi_ps_control --cov-report=term-missing"

# ---------------------------------------------------------------------------
# python-semantic-release
# ---------------------------------------------------------------------------
[tool.semantic_release]
# Version is managed exclusively through annotated git tags (e.g. v1.2.3).
# hatch-vcs reads the same tags at build time, so no version string lives
# anywhere in the source tree.
version_source = "tag"
tag_format = "v{version}"
commit_message = "chore(release): v{version} [skip ci]"
major_on_zero = false          # 0.x bumps: breaking → minor, feat/fix → patch

[tool.semantic_release.branches.main]
match = "main"
prerelease = false

[tool.semantic_release.commit_parser_options]
# Conventional Commits → SemVer mapping
# feat:         minor bump
# fix / perf:   patch bump
# BREAKING CHANGE or '!': major bump
allowed_tags = ["feat", "fix", "perf", "refactor", "docs", "style", "test", "chore", "ci", "build", "revert"]
minor_tags   = ["feat"]
patch_tags   = ["fix", "perf"]

[tool.semantic_release.changelog]
exclude_commit_patterns = [
  "^chore\\(release\\)",   # skip the release commit itself
  "^chore",
  "^ci",
  "^style",
  "^test",
]
